requirements.txt
requests
dnspython
beautifulsoup4
python-whois
